Added devel/dump_ay_addons.rb
